Fix ordering bugs in the toy grammar!
We need to check for if_statement before checking for assignment and
expr, because 'if' is also an identifier (we don't have reserved words
yet).  This affects the grammar shown on Medium.

Similarly, we need to check for multiple statements before a single
statement (this doesn't affect the Medium posts because the toy
grammar there starts at 'statement').