By default, compile to bytecode rather than to AST
Use mode=1 to return an AST, or mode=0 to return None.

Note that mode=1 is actually the slowest, since it has to convert the
internal AST to an ast.AST tree. mode=0 is the fastest (it creates an
internal AST and then throws it away).  mode=2 is the most realistic,
since this is what import does when loading a module for which no .pyc
file exists.

This reveals that compiling data/xxl.txt this way is actually faster
than the stdlib compile() function!

make time_stdlib: 1.36 real, 1.09 user, 0.26 sys (median of three)
make time:        0.90 real, 0.68 user, 0.20 sys (median of three)

Also, a run of `make simpy_cpython` now says:

Checked 1,641 files, 749,570 lines, 27,622,497 bytes in 6.700 seconds.
That's 111,883 lines/sec, or 4,123,014 bytes/sec.