Added grammar_grapher.py (#148)
This is an experiment in trying to visualize the overall structure of a grammar. The goal was to quickly get an idea of a grammar's complexity, and in particular, glean some insight into its "recursiveness". I'm not sure that I accomplished that goal, or that graphviz is the right tool for the job, but the resulting graphs are, at least, interesting and hopefully of some use.

I think with more refinement the output graphs can be made more useful (maybe with some colors and better labels, possibly some animation), but this seems like a good checkpoint for sharing what I have so far.