Fix segfault in debug mode when raising syntax errors
In some situations, we need to raise a syntax error with a token that
has an unitialized vale for col_offset (-1). This can happen with
invalid indentation errors. As we use that col_offset to get the source
for the traceback, we cannot use these unitialized values in those cases.