Fix segfault in debug mode when raising syntax errors (#204)
In some situations, we need to raise a syntax error with a token that
has an uninitialized value for col_offset (-1). This can happen with
invalid indentation errors. As we use that col_offset to get the source
for the traceback, we cannot use these uninitialized values in those cases.

Co-authored-by: Guido van Rossum <guido@python.org>